+++ src/main/java/com/github/erdragh/per_aspera/items/armour/ThrusterBootsMaterial.java
@@ -1,0 +1,10 @@
+package com.github.erdragh.per_aspera.items.armour;
+
+import com.github.alexnijjar.ad_astra.items.armour.SpaceSuitMaterial;
+
+public class ThrusterBootsMaterial extends SpaceSuitMaterial {
+    @Override
+    public String getName() {
+        return "thruster_boots";
+    }
+}