+++ src/main/java/com/github/erdragh/per_aspera/mixin/ItemRegistryMixin.java
@@ -3,9 +3,9 @@ package com.github.erdragh.per_aspera.mixin;
 import com.github.alexnijjar.ad_astra.items.armour.JetSuit;
 import com.github.alexnijjar.ad_astra.registry.ModItems;
 import com.github.erdragh.per_aspera.items.armour.ImprovedJetSuit;
-import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.item.ArmorMaterial;
-import net.minecraft.world.item.Item;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.item.ArmorMaterial;
+import net.minecraft.item.Item;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
@@ -28,9 +28,10 @@ public class ItemRegistryMixin {
                     target = "Lcom/github/alexnijjar/ad_astra/items/armour/JetSuit;*",
                     ordinal = 0
             ),
-            method = "<clinit>"
+            method = "<clinit>",
+            remap = false
     )
-    private static JetSuit jetSuitHelmet(ArmorMaterial material, EquipmentSlot slot, Item.Properties settings) {
+    private static JetSuit jetSuitHelmet(ArmorMaterial material, EquipmentSlot slot, Item.Settings settings) {
         return new ImprovedJetSuit(material, slot, settings);
     }
 
@@ -49,9 +50,10 @@ public class ItemRegistryMixin {
                     target = "Lcom/github/alexnijjar/ad_astra/items/armour/JetSuit;*",
                     ordinal = 0
             ),
-            method = "<clinit>"
+            method = "<clinit>",
+            remap = false
     )
-    private static JetSuit jetSuit(ArmorMaterial material, EquipmentSlot slot, Item.Properties settings) {
+    private static JetSuit jetSuit(ArmorMaterial material, EquipmentSlot slot, Item.Settings settings) {
         return new ImprovedJetSuit(material, slot, settings);
     }
 
@@ -70,9 +72,10 @@ public class ItemRegistryMixin {
                     target = "Lcom/github/alexnijjar/ad_astra/items/armour/JetSuit;*",
                     ordinal = 0
             ),
-            method = "<clinit>"
+            method = "<clinit>",
+            remap = false
     )
-    private static JetSuit jetSuitPants(ArmorMaterial material, EquipmentSlot slot, Item.Properties settings) {
+    private static JetSuit jetSuitPants(ArmorMaterial material, EquipmentSlot slot, Item.Settings settings) {
         return new ImprovedJetSuit(material, slot, settings);
     }
 
@@ -91,9 +94,10 @@ public class ItemRegistryMixin {
                     target = "Lcom/github/alexnijjar/ad_astra/items/armour/JetSuit;*",
                     ordinal = 0
             ),
-            method = "<clinit>"
+            method = "<clinit>",
+            remap = false
     )
-    private static JetSuit jetSuitBoots(ArmorMaterial material, EquipmentSlot slot, Item.Properties settings) {
+    private static JetSuit jetSuitBoots(ArmorMaterial material, EquipmentSlot slot, Item.Settings settings) {
         return new ImprovedJetSuit(material, slot, settings);
     }
 }